+++ GeorgeStore/Features/Carts/CartDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace GeorgeStore.Features.Carts;
 
 
-public sealed record CartItemDto(int Id, int ProductId, string Name, float Price, int Quantity, string Description, string Image)
+public sealed record CartItemDto(int Id, int ProductId, string Name, decimal Price, int Quantity, string Description, string Image)
 {
     public static CartItemDto FromEntity(CartItem CartItem)
     {
@@ -16,7 +16,7 @@ public sealed record CartItemDto(int Id, int ProductId, string Name, float Price
         );
     }
 }
-public sealed record CartDto(int Id, List<CartItemDto> Items, float Total);
+public sealed record CartDto(int Id, List<CartItemDto> Items, decimal Total);
 public sealed record AddItemRequest(int ProductId, int Quantity);
 
 public sealed record DecreaseItemDto(int ProductId);
