+++ GeorgeStore/Features/Products/ProductDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace GeorgeStore.Features.Products;
 
-public record ProductDto(int Id, string Name, float Price, string Description, string Image, int CategoryId, string CategoryName)
+public record ProductDto(int Id, string Name, decimal Price, string Description, string Image, int CategoryId, string CategoryName)
 {
     public static ProductDto FromEntity(Product entity)
     {
@@ -8,4 +8,4 @@ public record ProductDto(int Id, string Name, float Price, string Description, s
     }
 }
 
-public record ProductCreateDTO(string Name, float Price, string Description, string Image, int CategoryId);
+public record ProductCreateDTO(string Name, decimal Price, string Description, string Image, int CategoryId);
